rename market_recommender backend and streamlit files to reflect their purpose:
market_recommender_api.py
market_recommender_page.py
